+++ lib/src/index.d.ts
@@ -1,4 +1,2 @@
 export * from './nameConverter';
-export declare const isRelationshipField: (fieldName: string) => boolean;
-import getSchemaFromData from './getSchemaFromData';
-export default getSchemaFromData;
+export * from './SchemaBuilder';