+++ lib/src/utils/TypeCheck.d.ts
@@ -1,0 +1,4 @@
+export declare function isObject(value: any): boolean;
+export declare function isDate(value: any): boolean;
+export declare function isFloat(value: string): boolean;
+export declare function every(values: any[], type: string | ((val: any) => boolean)): boolean;