+++ src/index.ts
@@ -1,6 +1,2 @@
 export * from './nameConverter';
-import isRelationshipFieldImport from './isRelationshipField';
-export const isRelationshipField = isRelationshipFieldImport;
-import getSchemaFromData from './getSchemaFromData';
-
-export default getSchemaFromData;
+export * from './SchemaBuilder'