+++ src/isRelationshipField.ts
@@ -1,1 +1,0 @@
-export default (fieldName: string) => fieldName.endsWith('_id');