+++ src/testDate.ts
@@ -1,0 +1,7 @@
+const fakeDate = new Date('1234-23123-asdad')
+
+console.log(Number.isInteger('2020-07-22'))
+console.log(fakeDate);
+console.log(fakeDate.getDate());
+console.log(typeof fakeDate);
+console.log(fakeDate instanceof Date)