+++ tests/SchemaBuilder.spec.ts
@@ -1,16 +1,6 @@
-import {
-    GraphQLBoolean,
-    GraphQLID,
-    // GraphQLInputObjectType,
-    // GraphQLInputType,
-    GraphQLInt,
-    GraphQLList,
-    // GraphQLNamedType,
-    GraphQLNonNull,
-    GraphQLObjectType,
-    GraphQLString,
-} from 'graphql';
-import getSchemaFromData from '../src/getSchemaFromData';
+import { GraphQLObjectType, GraphQLNonNull, GraphQLID, GraphQLString, GraphQLInt, GraphQLList, GraphQLBoolean, GraphQLFloat } from 'graphql';
+import { SchemaBuilder } from '../src/'
+import DateType from '../src/DateType';
 
 const data = {
     posts: [
@@ -18,13 +8,31 @@ const data = {
             id: 1,
             title: 'Lorem Ipsum',
             views: 254,
+            created_at: new Date("2020-02-22"),
+            rating: 4.5,
             user_id: 123,
+            hidden: true,
+            audits: [
+                {
+                    user: "Admin",
+                    message: "Lorem Ipsum"
+                }
+            ]
         },
         {
             id: 2,
             title: 'Sic Dolor amet',
             views: 65,
+            created_at: new Date("2020-03-05"),
+            rating: 3.1,
             user_id: 456,
+            hidden: false,
+            audits: [
+                {
+                    user: "Admin",
+                    message: "Sic Dolor amet"
+                }
+            ]
         },
     ],
     users: [
@@ -39,13 +47,70 @@ const data = {
     ],
 };
 
+const dataWithKeys = {
+    posts: [
+        {
+            PostID: 1,
+            title: 'Lorem Ipsum',
+            views: 254,
+            created_at: new Date("2020-02-22"),
+            rating: 4.5,
+            UserID: 123,
+            hidden: true,
+            audits: [
+                {
+                    user: "Admin",
+                    message: "Lorem Ipsum"
+                }
+            ]
+        },
+        {
+            PostID: 2,
+            title: 'Sic Dolor amet',
+            views: 65,
+            created_at: new Date("2020-03-05"),
+            rating: 3.1,
+            UserID: 456,
+            hidden: false,
+            audits: [
+                {
+                    user: "Admin",
+                    message: "Sic Dolor amet"
+                }
+            ]
+        },
+    ],
+    users: [
+        {
+            UserID: 123,
+            name: 'John Doe',
+        },
+        {
+            UserID: 456,
+            name: 'Jane Doe',
+        },
+    ],
+}
+
+const AuditType: GraphQLObjectType = new GraphQLObjectType({
+    name: 'Audit',
+    fields: () => ({
+        user: { type: new GraphQLNonNull(GraphQLString) },
+        message: { type: new GraphQLNonNull(GraphQLString) },
+    }),
+});
+
 const PostType: GraphQLObjectType = new GraphQLObjectType({
     name: 'Post',
     fields: () => ({
         id: { type: new GraphQLNonNull(GraphQLID) },
         title: { type: new GraphQLNonNull(GraphQLString) },
         views: { type: new GraphQLNonNull(GraphQLInt) },
+        created_at: { type: new GraphQLNonNull(DateType) },
+        rating: { type: new GraphQLNonNull(GraphQLFloat) },
         user_id: { type: new GraphQLNonNull(GraphQLID) },
+        hidden: { type: new GraphQLNonNull(GraphQLBoolean) },
+        audits: { type: new GraphQLList(AuditType) },
         User: { type: UserType },
     }),
 });
@@ -59,67 +124,73 @@ const UserType: GraphQLObjectType = new GraphQLObjectType({
     }),
 });
 
-/*
-const ListMetadataType = new GraphQLObjectType({
-    name: 'ListMetadata',
-    fields: {
-        count: { type: GraphQLInt },
-    },
+const PostTypeCustom: GraphQLObjectType = new GraphQLObjectType({
+    name: 'Post',
+    fields: () => ({
+        PostID: { type: new GraphQLNonNull(GraphQLID) },
+        title: { type: new GraphQLNonNull(GraphQLString) },
+        views: { type: new GraphQLNonNull(GraphQLInt) },
+        created_at: { type: new GraphQLNonNull(DateType) },
+        rating: { type: new GraphQLNonNull(GraphQLFloat) },
+        UserID: { type: new GraphQLNonNull(GraphQLID) },
+        hidden: { type: new GraphQLNonNull(GraphQLBoolean) },
+        audits: { type: new GraphQLList(AuditType) },
+        User: { type: UserType },
+    }),
 });
 
-const QueryType = new GraphQLObjectType({
-    name: 'Query',
-    fields: {
-        getPost: {
-            type: PostType,
-            args: {
-                id: { type: new GraphQLNonNull(GraphQLID) },
-            },
-        },
-        getPageOfPost: {
-            type: new GraphQLList(PostType),
-            args: {
-                page: { type: GraphQLInt },
-                perPage: { type: GraphQLInt },
-                sortField: { type: GraphQLString },
-                sortOrder: { type: GraphQLString },
-                filter: { type: GraphQLString },
-            },
-        },
-        getUser: {
-            type: UserType,
-            args: {
-                id: { type: new GraphQLNonNull(GraphQLID) },
-            },
-        },
-        getPageOfUser: {
-            type: new GraphQLList(UserType),
-            args: {
-                page: { type: GraphQLInt },
-                perPage: { type: GraphQLInt },
-                sortField: { type: GraphQLString },
-                sortOrder: { type: GraphQLString },
-                filter: { type: GraphQLString },
-            },
-        },
-    },
+const UserTypeCustom: GraphQLObjectType = new GraphQLObjectType({
+    name: 'User',
+    fields: () => ({
+        UserID: { type: new GraphQLNonNull(GraphQLID) },
+        name: { type: new GraphQLNonNull(GraphQLString) },
+        Posts: { type: new GraphQLList(PostType) },
+    }),
 });
-*/
-describe('getSchemaFromData ', () => {
-    const schema = getSchemaFromData(data)
-    const typeMap = schema.getTypeMap()
+
+
+
+describe("SchemaBuild ", () => {
+    describe("Without user provided primary keys", () => {
+        createTest(data, PostType, UserType)
+    })
+    describe("With user provided primary keys", () => {
+        createTest(dataWithKeys, PostTypeCustom, UserTypeCustom, "users:UserID,posts:PostID")
+    })
+})
+describe('Inflection ', () => {
+    test('pluralizes and capitalizes correctly', () => {
+        const data = {
+            feet: [{ id: 1, size: 42 }, { id: 2, size: 39 }],
+            categories: [{ id: 1, name: 'foo' }],
+        };
+        const schema = new SchemaBuilder(data).schema
+        const queries = schema.getQueryType()!.getFields();
+        expect(queries).toHaveProperty('Foot');
+        expect(queries).toHaveProperty('Category');
+        expect(queries).toHaveProperty('allFeet');
+        expect(queries).toHaveProperty('allCategories');
+        const types = schema.getTypeMap();
+        expect(types).toHaveProperty('Foot');
+        expect(types).toHaveProperty('Category');
+    });
+})
+
+function createTest(data: Record<string, any>, posts: GraphQLObjectType, users: GraphQLObjectType, pkeys = "") {
+    const schema = new SchemaBuilder(data, pkeys)
+    const typeMap = schema.schema.getTypeMap()
     const postTypeMap = typeMap['Post'] as GraphQLObjectType
     const userTypeMap = typeMap['User'] as GraphQLObjectType
     test('creates one type per data type', () => {
         // expect(postTypeMap.getFields()).toBeDefined()
         expect(postTypeMap!.name).toBeDefined()
-        expect(postTypeMap!.name).toEqual(PostType.name);
+        expect(postTypeMap!.name).toEqual(posts.name);
         expect(Object.keys(postTypeMap.getFields())).toEqual(
-            Object.keys(PostType.getFields())
+            Object.keys(posts.getFields())
         );
-        expect(userTypeMap!.name).toEqual(UserType.name);
+        expect(userTypeMap!.name).toEqual(users.name);
         expect(Object.keys(userTypeMap.getFields())).toEqual(
-            Object.keys(UserType.getFields())
+            Object.keys(users.getFields())
         );
     });
 
@@ -132,8 +203,8 @@ describe('getSchemaFromData ', () => {
     });
 
     test('creates seven query fields per data type', () => {
-        const queries = schema!.getQueryType()!.getFields();
-        expect((queries['Post'].type as GraphQLObjectType).name).toEqual(PostType.name);
+        const queries = schema.schema.getQueryType()!.getFields();
+        expect((queries['Post'].type as GraphQLObjectType).name).toEqual(posts.name);
         expect(queries['Post'].args).toEqual([
             {
                 astNode: undefined,
@@ -141,7 +212,7 @@ describe('getSchemaFromData ', () => {
                 deprecationReason: undefined,
                 description: undefined,
                 extensions: undefined,
-                name: 'id',
+                name: queries['Post'].args[0].name,
                 type: new GraphQLNonNull(GraphQLID),
             },
         ]);
@@ -158,7 +229,7 @@ describe('getSchemaFromData ', () => {
         expect(queries['allPosts'].args[4].type.toString()).toEqual('PostFilter');
         expect(queries['_allPostsMeta'].type.toString()).toEqual('ListMetadata');
 
-        expect((queries['User'].type as GraphQLObjectType).name).toEqual(UserType.name);
+        expect((queries['User'].type as GraphQLObjectType).name).toEqual(users.name);
         expect(queries['User'].args).toEqual([
             {
                 astNode: undefined,
@@ -166,7 +237,7 @@ describe('getSchemaFromData ', () => {
                 deprecationReason: undefined,
                 description: undefined,
                 extensions: undefined,
-                name: 'id',
+                name: queries['User'].args[0].name,
                 type: new GraphQLNonNull(GraphQLID),
             },
         ]);
@@ -185,90 +256,16 @@ describe('getSchemaFromData ', () => {
     });
 
     test('creates seven mutation fields per data type', () => {
-        const mutations = schema.getMutationType()!.getFields();
-        expect((mutations['createPost'].type as GraphQLObjectType).name).toEqual(PostType.name);
-        expect(mutations['createPost'].args).toEqual([
-            {
-                astNode: undefined,
-                name: 'id',
-                type: new GraphQLNonNull(GraphQLID),
-                defaultValue: undefined,
-                deprecationReason: undefined,
-                description: undefined,
-                extensions: undefined,
-            },
-            {
-                astNode: undefined,
-                name: 'title',
-                type: new GraphQLNonNull(GraphQLString),
-                defaultValue: undefined,
-                deprecationReason: undefined,
-                description: undefined,
-                extensions: undefined,
-            },
-            {
-                astNode: undefined,
-                name: 'views',
-                type: new GraphQLNonNull(GraphQLInt),
-                defaultValue: undefined,
-                deprecationReason: undefined,
-                description: undefined,
-                extensions: undefined,
-            },
-            {
-                astNode: undefined,
-                name: 'user_id',
-                type: new GraphQLNonNull(GraphQLID),
-                defaultValue: undefined,
-                deprecationReason: undefined,
-                description: undefined,
-                extensions: undefined,
-            },
-        ]);
-        expect((mutations['updatePost'].type as GraphQLObjectType).name).toEqual(PostType.name);
-        expect(mutations['updatePost'].args).toEqual([
-            {
-                astNode: undefined,
-                name: 'id',
-                type: new GraphQLNonNull(GraphQLID),
-                defaultValue: undefined,
-                deprecationReason: undefined,
-                description: undefined,
-                extensions: undefined,
-            },
-            {
-                astNode: undefined,
-                name: 'title',
-                type: GraphQLString,
-                defaultValue: undefined,
-                deprecationReason: undefined,
-                description: undefined,
-                extensions: undefined,
-            },
-            {
-                astNode: undefined,
-                name: 'views',
-                type: GraphQLInt,
-                defaultValue: undefined,
-                deprecationReason: undefined,
-                description: undefined,
-                extensions: undefined,
-            },
-            {
-                astNode: undefined,
-                name: 'user_id',
-                type: GraphQLID,
-                defaultValue: undefined,
-                deprecationReason: undefined,
-                description: undefined,
-                extensions: undefined,
-            },
-        ]);
+        const mutations = schema.schema.getMutationType()!.getFields();
+        expect((mutations['createPost'].type as GraphQLObjectType).name).toEqual(posts.name);
+        expect(mutations['createPost'].args.length).toEqual(8)
+        expect((mutations['updatePost'].type as GraphQLObjectType).name).toEqual(posts.name);
+        expect(mutations['updatePost'].args.length).toEqual(8);
         expect((mutations['removePost'].type as typeof GraphQLBoolean).name).toEqual(GraphQLBoolean.name);
         expect(mutations['removePost'].args).toEqual([
             {
                 astNode: undefined,
-                name: 'id',
+                name: mutations['removePost'].args[0].name,
                 type: new GraphQLNonNull(GraphQLID),
                 defaultValue: undefined,
                 deprecationReason: undefined,
@@ -276,11 +273,11 @@ describe('getSchemaFromData ', () => {
                 extensions: undefined,
             },
         ]);
-        expect((mutations['createUser'].type as GraphQLObjectType).name).toEqual(UserType.name);
+        expect((mutations['createUser'].type as GraphQLObjectType).name).toEqual(users.name);
         expect(mutations['createUser'].args).toEqual([
             {
                 astNode: undefined,
-                name: 'id',
+                name: mutations['createUser'].args[0].name,
                 type: new GraphQLNonNull(GraphQLID),
                 defaultValue: undefined,
                 deprecationReason: undefined,
@@ -297,11 +294,11 @@ describe('getSchemaFromData ', () => {
                 extensions: undefined,
             },
         ]);
-        expect((mutations['updateUser'].type as GraphQLObjectType).name).toEqual(UserType.name);
+        expect((mutations['updateUser'].type as GraphQLObjectType).name).toEqual(users.name);
         expect(mutations['updateUser'].args).toEqual([
             {
                 astNode: undefined,
-                name: 'id',
+                name: mutations['updateUser'].args[0].name,
                 type: new GraphQLNonNull(GraphQLID),
                 defaultValue: undefined,
                 deprecationReason: undefined,
@@ -326,26 +323,9 @@ describe('getSchemaFromData ', () => {
                 deprecationReason: undefined,
                 description: undefined,
                 extensions: undefined,
-                name: 'id',
+                name: mutations['removeUser'].args[0].name,
                 type: new GraphQLNonNull(GraphQLID),
             },
         ]);
     });
-})
-describe('Inflection ', () => {
-    test('pluralizes and capitalizes correctly', () => {
-        const data = {
-            feet: [{ id: 1, size: 42 }, { id: 2, size: 39 }],
-            categories: [{ id: 1, name: 'foo' }],
-        };
-        const queries = getSchemaFromData(data).getQueryType()!.getFields();
-        expect(queries).toHaveProperty('Foot');
-        expect(queries).toHaveProperty('Category');
-        expect(queries).toHaveProperty('allFeet');
-        expect(queries).toHaveProperty('allCategories');
-        const types = getSchemaFromData(data).getTypeMap();
-        expect(types).toHaveProperty('Foot');
-        expect(types).toHaveProperty('Category');
-    });
-})
-
+}